pyproject.toml
[project]
name = "to-do"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asqlite>=2.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.120.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "backend/src"  # <-- This is the magic line
]
asyncio_mode = "auto"
